requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
contextlib2==21.6.0
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-ninja==1.3.0
django-ninja-extra==0.21.4
django-ninja-jwt==5.3.3
dnspython==2.6.1
email_validator==2.2.0
idna==3.8
injector==0.22.0
isodate==0.6.1
libpass==1.7.5.post0
oauthlib==3.2.2
passlib==1.7.4
portalocker==2.10.1
psycopg==3.2.2
psycopg-binary==3.2.2
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
python-dotenv==1.0.1
pywin32==306
requests==2.32.3
requests-oauthlib==2.0.0
ruff==0.6.7
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
